change "home" link in logout page
Originally there was a link called "home" that took the user back to the admin login page. But there is also a link called "log in again" that does the same thing. So I changed the "home" link to "Visitor Management Home", and instead of taking the user back to the admin login it takes them to the VMS homepage.